+++ src/view/secondary-bar/collapsed-sidebar-items.tsx
